Add files via upload
Arduino starts to tweak if you hold down the keys (probably because data is being sent to the serial monitor faster than the Arduino code is reading from it and as a result, the messages being sent over appear as one long string with more than 6 angle inputs rather than the normal one. Maybe change the Arduino code to add a new Serial.readUntil() that will fix this bug?). To control it, tap the keys (the code should have comments on what each key does and what it turns). Also, remember that I have the baud rate set in this file as 38400; Make sure to change it back to 9600 or set the Arduino baud rate in Serial.begin() to 38400.